feat: add method createFloatScLink

+++ src/main/java/org/jmantic/scmemory/websocketmemory/sync/ScLinkImpl.java
@@ -47,7 +47,7 @@ class ScLinkFloatImpl implements ScLinkFloat {
     }
 
     @JsonIgnore
-    public void setContent(int content) {
+    public void setContent(float content) {
         this.content = content;
     }
